Seperate file for Command - ALEKSANDR VINOKHODOV CHECK
