+++ src/CommandProcessing.h
@@ -4,29 +4,11 @@
 #include <string.h>
 #include <fstream>
 #include <list>
+#include "Command.h"
 #include "GameEngine.h"
 using namespace std;
 
 string strToLower(string str);
-class Command
-{
-private:
-    string command;
-    string effect;
-
-public:
-    Command();                            // default
-    Command(string com, string eff);      // parametrized constr
-    Command(const Command &c);            // copy constr
-    Command &operator=(const Command &c); // assignment operator overload
-    ~Command();                           // destructor
-    void saveEffect(string eff);
-    string getCommand() { return this->command; };
-    string getEffect() { return effect; };
-    void setCommand(string newCommand) { command = newCommand; };
-    void setEffect(string newEffect) { effect = newEffect; };
-    friend ostream &operator<<(ostream &os, Command &c);
-};
 
 class CommandProcessor
 {
